back/requirements.txt
Django==6.0.4
djangorestframework==3.17.1
channels==4.3.2
channels-redis==4.3.0
daphne==4.2.1
django-cors-headers==4.9.0
dj-database-url==3.1.2
psycopg[binary]==3.3.4
python-dotenv==1.2.2
whitenoise==6.9.0
